% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{anova_2x2_table}
\alias{anova_2x2_table}
\title{Title}
\usage{
anova_2x2_table(
  factor_a_name = "Factor A",
  factor_b_name = "Factor B",
  factor_a_levels = c("\\\\(A_1\\\\)", "\\\\(A_2\\\\)"),
  factor_b_levels = c("\\\\(B_1\\\\)", "\\\\(B_2\\\\)"),
  cell_values = c("\\\\(A_1 B_1\\\\)", "\\\\(A_2 B_1\\\\)", "\\\\(A_2 B_1\\\\)",
    "\\\\(A_2 B_2\\\\)")
)
}
\arguments{
\item{cell_values}{}
}
\value{

}
\description{
Title
}
