% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_beam.R
\name{balance_beam}
\alias{balance_beam}
\title{Balance beam plot}
\usage{
balance_beam(data, label_boxes = FALSE, size = 4)
}
\arguments{
\item{data}{A vector of integers.}

\item{label_boxes}{Logical. Show labels in boxes with their deviation scores.}

\item{size}{The size of text labels.}
}
\value{
A ggplot2 plot showing a box histogram of scores perched on a
triangle positioned at the mean of the data
}
\description{
Balance beam plot
}
\examples{
balance_beam(c(1, 2, 6, 6, 10))
}
