% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{frequency_table}
\alias{frequency_table}
\title{Make a frequency table}
\usage{
frequency_table(
  x,
  f = NULL,
  values_name = "\\\\(X\\\\)",
  binwidth = 1,
  additional_cols = FALSE,
  digits = 2
)
}
\arguments{
\item{additional_cols}{}
}
\value{

}
\description{
Make a frequency table
}
