% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-utils.R
\name{input_slider}
\alias{input_slider}
\title{Title}
\usage{
input_slider(
  name,
  label,
  value,
  min,
  max,
  step = 1,
  width = "50\%",
  scale_name = "scale"
)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
