% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert.R
\name{likert}
\alias{likert}
\title{HTML Likert scale}
\usage{
likert(question, response_options, scale_name = "likert", extra_css = "")
}
\arguments{
\item{extra_css}{}
}
\value{

}
\description{
HTML Likert scale
}
