% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_correlation_generic}
\alias{plot_correlation_generic}
\title{Title}
\usage{
plot_correlation_generic(df, ellipse = FALSE, line = FALSE, ...)
}
\arguments{
\item{line}{}
}
\value{

}
\description{
Title
}
