% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Title}
\usage{
plot_distribution(
  fun = dnorm,
  args = NULL,
  xlim = NULL,
  xbreaks = NULL,
  xlab = NULL,
  segments = NULL,
  linetype = 1,
  fill_colors = NULL,
  alpha = 0.7,
  arrow = NULL,
  arrow_y = 0.4,
  arrow_size = 0.2
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Title
}
