% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Title}
\usage{
plot_histogram(
  x,
  binwidth = 1,
  breaks = NULL,
  xlab = "Score",
  ylab = "Frequency",
  block = FALSE
)
}
\arguments{
\item{block}{}
}
\value{

}
\description{
Title
}
