% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nhst_diagram.R
\name{plot_nhst_diagram}
\alias{plot_nhst_diagram}
\title{Title}
\usage{
plot_nhst_diagram(
  text.size = 7,
  pop1lab = "Known\\nOriginal\\npopulation",
  pop1color = "black",
  pop1line = 1,
  pop2lab = "Unknown\\nTreated\\npopulation",
  pop2color = "black",
  pop2fill = "lightskyblue",
  pop2line = 2,
  sample1lab = "Sample",
  sample1color = "black",
  sample1line = 1,
  sample2lab = "Treated\\nSample",
  sample2color = "black",
  sample2fill = "lightskyblue",
  sample2line = 1,
  sample_line = "black"
)
}
\arguments{
\item{sample_line}{}
}
\value{

}
\description{
Title
}
