% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_save}
\alias{plot_save}
\title{Title}
\usage{
plot_save(
  filename,
  plot = last_plot(),
  dir = "slides",
  subdir = "media",
  width = 5,
  height = 3
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
Title
}
