% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability-tables.R
\name{SD_table}
\alias{SD_table}
\title{Standard deviation table}
\usage{
SD_table(
  x,
  population_sd = FALSE,
  include = c("mean", "devs", "devs_sq", "ss", "var", "sd"),
  digits = 2,
  align = "r",
  x_var_name = "\\\\(X\\\\)"
)
}
\arguments{
\item{x_var_name}{}
}
\value{

}
\description{
Standard deviation table
}
