% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-tables.R
\name{t_table}
\alias{t_table}
\title{t table}
\usage{
t_table(
  dfs = 1:15,
  alphas = c(0.1, 0.05, 0.025, 0.01, 0.005),
  two_tailed_only = FALSE,
  dots = length(dfs)
)
}
\arguments{
\item{dots}{}
}
\value{

}
\description{
t table
}
